Normal (drawing) mode from current implementation.
