+++ Main.hs
@@ -4,6 +4,8 @@ import Dhek.Engine (makeInterpreter)
 import Dhek.GUI (makeGUI, runGUI)
 import Dhek.Signal (connectSignals)
 
+import Dhek.Mode.Normal
+
 main :: IO ()
 main = do
     gui <- makeGUI
